BuildMaster release: v1.0.1.26

+++ Pod/Classes/Terminals/HPA/Models/ParserObject/ResponseObjects/HpaFileUpload.m
@@ -11,7 +11,7 @@
         NSData *data = [NSData dataWithContentsOfURL:url];
         self.fileName = url.lastPathComponent;
         self.hexData = [self dataTohex:data];
-        self.fileSize = [NSString stringWithFormat:@"%ld",data.length];
+        self.fileSize = [NSString stringWithFormat:@"%lu",(unsigned long)data.length];
     }
     return self;
 }
